separate log_value_error decorators for dynesty and ultranest as they raise different exceptions
